+++ common/config.go
@@ -5,21 +5,12 @@ import (
 
 	"github.com/agglayer/aggkit/config/types"
 	ethermanconfig "github.com/agglayer/aggkit/etherman/config"
-	"github.com/agglayer/aggkit/translator"
 )
 
 // Config holds the configuration for the Aggkit.
 type Config struct {
-	// TODO: @Stefan-Ethernal DEPRECATE
-	// IsValidiumMode has the value true if the sequence sender is running in validium mode.
-	IsValidiumMode bool `mapstructure:"IsValidiumMode"`
 	// NetworkID is the networkID of the Aggkit being run
 	NetworkID uint32 `mapstructure:"NetworkID"`
-	// TODO: @Stefan-Ethernal DEPRECATE
-	// Contract Versions: elderberry, banana
-	ContractVersions string `mapstructure:"ContractVersions"`
-	// TODO: @Stefan-Ethernal DEPRECATE
-	Translator translator.Config `mapstructure:"Translator"`
 	// L2URL is the URL of the L2 node
 	L2RPC ethermanconfig.RPCClientConfig `mapstructure:"L2RPC"`
 }
