+++ etherman/config/config.go
@@ -1,25 +1,13 @@
 package config
 
 import (
-	"github.com/0xPolygon/zkevm-ethtx-manager/etherman"
 	"github.com/ethereum/go-ethereum/common"
 )
 
-// Config represents the configuration of the etherman
-type Config struct {
-	// URL is the URL of the Ethereum node for L1
-	URL string `mapstructure:"URL"`
-
-	// TODO: @Stefan-Ethernal Deprecate
-	EthermanConfig etherman.Config
-
-	// TODO: @Stefan-Ethernal Deprecate
-	// ForkIDChunkSize is the max interval for each call to L1 provider to get the forkIDs
-	ForkIDChunkSize uint64 `mapstructure:"ForkIDChunkSize"`
-}
-
 // L1Config represents the configuration of the network used in L1
 type L1Config struct {
+	// URL is the URL of the Ethereum node for L1
+	URL string `mapstructure:"URL"`
 	// Chain ID of the L1 network
 	L1ChainID uint64 `json:"chainId"`
 	// ZkEVMAddr Address of the L1 contract polygonZkEVMAddress
