--- translator/translator.go
@@ -1,5 +1,0 @@
-package translator
-
-type Translator interface {
-	Translate(contextName string, data string) string
-}
